progress/requirements.txt
alembic==1.4.2
appdirs==1.4.3
CacheControl==0.12.6
cachetools==4.1.0
certifi==2019.11.28
chardet==3.0.4
click==7.1.2
colorama==0.4.3
contextlib2==0.6.0
distlib==0.3.0
distro==1.5.0
dnspython==1.16.0
eventlet==0.25.2
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Login==0.5.0
Flask-Migrate==2.5.3
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
gevent==20.5.0
gevent-websocket==0.10.1
greenlet==0.4.15
gunicorn==20.0.4
html5lib==1.0.1
idna==2.9
ipaddr==2.2.0
itsdangerous==1.1.0
Jinja2==2.11.2
lockfile==0.12.2
Mako==1.1.3
MarkupSafe==1.1.1
monotonic==1.5
msgpack==0.6.2
packaging==20.3
pep517==0.8.2
progress==1.5
psycopg2-binary==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.13.0
python-editor==1.0.4
pytoml==0.1.21
requests==2.23.0
requests-oauthlib==1.3.0
retrying==1.3.3
rsa==4.0
six==1.14.0
SQLAlchemy==1.3.17
urllib3==1.25.8
webencodings==0.5.1
Werkzeug==1.0.1
